Fix bug when save json file of predicted results in vis.py

+++ lib/utils/vis.py
@@ -528,7 +528,7 @@ def write_pose_to_json(im_name, output_dir, boxes, car_cls_prob, euler_angle, tr
             # filter out by ignored_mask_binary
             car_info = dict()
             car_info["car_id"] = int(car_model_i)
-            car_info["box"] = [round(x) for x in box_i]
+            car_info["box"] = [int(round(x)) for x in box_i]
             car_info["pose"] = [float(x) for x in euler_angle_i] + [float(x) for x in trans_pred_i]
             # We use rectangle area
             car_info["area"] = int(areas[i])
@@ -547,5 +547,3 @@ def write_pose_to_json(im_name, output_dir, boxes, car_cls_prob, euler_angle, tr
 
     with open(json_file, 'w') as outfile:
         json.dump(car_list, outfile, indent=4)
-
-
